Ayush

+++ src/utils/cloudinary.js
@@ -28,6 +28,26 @@ const uploadOnCloudinary = async (localFilePath) => {
         //file has been uploaded successfully
         // console.log("file uploaded on cloudinary", responce.url);
         //
+        await fs.unlinkSync(localFilePath); // unlik file path
+        return responce; // for user
+    } catch (error) {
+        console.log("error on cloudinary uploading :", error);
+        await fs.unlinkSync(localFilePath); // will remove local saved file if upload file failed
+        return null;
+    }
+};
+
+const uploadOnCloudinaryPDF = async (localFilePath) => {
+    try {
+        if (!localFilePath) return null;
+        // ** upload file on cloudinary
+        const responce = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "raw",
+            format: "pdf",
+        });
+        //file has been uploaded successfully
+        // console.log("file uploaded on cloudinary", responce.url);
+        //
         fs.unlinkSync(localFilePath); // unlik file path
         return responce; // for user
     } catch (error) {
@@ -74,4 +94,4 @@ const deleteImageOnCloudinary = async (imageUrl) => {
         return null;
     }
 };
-export { uploadOnCloudinary, deleteImageOnCloudinary };
+export { uploadOnCloudinary, deleteImageOnCloudinary, uploadOnCloudinaryPDF };
